refactor : 소수점 아래 계산 가능하도록 수정

+++ src/main/kotlin/com/nextstep/jngcii/step2/Computer.kt
@@ -3,7 +3,7 @@ package com.nextstep.jngcii.step2
 class Computer(
     private val calculator: Calculator
 ) {
-    fun compute(requests: Requests): Int {
+    fun compute(requests: Requests): Double {
         var left = requests.numbers.first()
 
         for (index in 0 until requests.operators.size) {
@@ -15,6 +15,6 @@ class Computer(
             left = calculator.run(left, right, operator)
         }
 
-        return left.value
+        return left.value.toDouble()
     }
 }
